refactor: refactor de estructura de proyecto y nombres de clases
