+++ app/src/main/java/com/ebcf/jnab/data/model/SymposiumModel.kt
@@ -2,7 +2,7 @@ package com.ebcf.jnab.data.model
 
 import java.time.LocalDateTime
 
-data class Symposium(
+data class SymposiumModel(
     val title: String,
     val description: String,
     val dateTime: LocalDateTime
