+++ app/src/main/java/com/ebcf/jnab/data/repository/SymposiumRepository.kt
@@ -2,7 +2,7 @@ package com.ebcf.jnab.data.repository
 
 import android.os.Build
 import androidx.annotation.RequiresApi
-import com.ebcf.jnab.data.model.Symposium
+import com.ebcf.jnab.data.model.SymposiumModel
 import java.time.LocalDateTime
 
 class SymposiumRepository {
@@ -10,17 +10,17 @@ class SymposiumRepository {
     // Lista de simposios hardcodeados
     @RequiresApi(Build.VERSION_CODES.O)
     private val hardcodedSymposia = listOf(
-        Symposium(
+        SymposiumModel(
             title = "Simposio de Antropología Biológica 2025",
             description = "Un simposio sobre avances en la investigación de la antropología biológica.",
             dateTime = LocalDateTime.of(2025, 5, 1, 9, 0)
         ),
-        Symposium(
+        SymposiumModel(
             title = "Simposio sobre Evolución Humana",
             description = "Simposio que cubre temas relacionados con la evolución humana.",
             dateTime = LocalDateTime.of(2024, 11, 15, 10, 0)
         ),
-        Symposium(
+        SymposiumModel(
             title = "Simposio de Biodiversidad Humana",
             description = "Un análisis sobre la biodiversidad en los seres humanos.",
             dateTime = LocalDateTime.of(2025, 7, 20, 14, 0)
@@ -28,7 +28,7 @@ class SymposiumRepository {
     )
 
     @RequiresApi(Build.VERSION_CODES.O)
-    fun getAll(): List<Symposium> {
+    fun getAll(): List<SymposiumModel> {
         return hardcodedSymposia
     }
