+++ app/src/main/res/layout/fragment_symposiums_list.xml
