+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">test</string>
+    <string name="app_name">JNAP</string>
     <string name="title_symposiums">Simposios</string>
     <string name="title_agenda">Agenda</string>
     <string name="title_speakers">Expositores</string>
